+++ mobile/android/android-components/buildSrc/src/main/java/Gecko.kt
@@ -6,7 +6,7 @@ internal object GeckoVersions {
     /**
      * GeckoView Nightly Version.
      */
-    const val nightly_version = "78.0.20200526023857"
+    const val nightly_version = "78.0.20200527093527"
 
     /**
      * GeckoView Beta Version.